Create README.md
Given two singly linked lists that intersect at some point, find the intersecting node. Assume the lists are non-cyclical.
For example, given A = 3 ➔ 7 ➔ 8 ➔ 10 and B = 99 ➔ 1 ➔ 8 ➔ 10, return the node with value 8. In this example, assume nodes with the same value are the exact same node objects.

Do this in 0( m + n) time (where m and n are the lengths of the lists) and constant space.

Implement a stack that has the following methods: • push ( v a 1 ) : push v a 1 onto the stack • pop: pop off and return the topmost element of the stack. If there are no elements in the stack, throw an error. • max: return the maximum value in the stack currently. If there are no elements in the stack, throw an error.
Each method should run in constant time.